fix readme and rename pages folder
