+++ templates/admin/src/components/App.tsx
@@ -7,8 +7,8 @@ import Router from './layout/Router';
 import Navigation from './navigation/Navigation';
 import SessionService from '../services/session/SessionService';
 import Header from './layout/Header';
-import Login from './pages/login/Login';
 import ConditionalRoute from './theme/routes/ConditionalRoute';
+import Login from './features/login/Login';
 import GlobalErrorBoundary from './theme/GlobalErrorBoundary';
 import NotificationRenderer from './theme/NotificationRenderer';
 import LocaleService from '../i18n/locale/LocaleService';
