+++ templates/admin/src/components/layout/Router.tsx
@@ -5,7 +5,7 @@ import {
   Redirect,
 } from 'react-router-dom';
 import { getGlobalInstance } from 'plume-ts-di';
-import Home from '../pages/Home';
+import Home from '../features/Home';
 import Users from '../../lib/plume-admin-users/pages/Users';
 import PermissionRoute from '../theme/routes/PermissionRoute';
 import Permission from '../../services/session/Permission';
